# Introduce BOP_CMP for optimized comparisons