--- a/benchmark/range_min.yml
+++ b/benchmark/range_min.yml
@@ -1,0 +1,2 @@
+benchmark:
+  - (1..10).min