This PR introduces a new basic operator, BOP_CMP, so we can quickly check whether `<=>` has been redefined when doing optimized comparisons.

The PR is in two commits (plus one more to update the deps), which could be reviewed separtely:

Move BOP macros to separate file
---

This first commit moves `ruby_basic_operators` and the `UNREDEFINED` macros out of
vm_core.h and into basic_operators.h so that we can use them in internal/compar.h, as well as more broadly in places where we currently do a method look up via `rb_method_basic_definition_p` (e.g. object.c, numeric.c, complex.c, enum.c, and a number of other places if we introduce more BOPs).

The most controversial part of this first commit is probably moving `redefined_flag` out of `rb_vm_t`. Neither [vm_opt_method_def_table or vm_opt_mid_table](https://github.com/ruby/ruby/blob/9da2a5204f32a4f2ce135fddde2abb6e07d647e9/vm.c) are part of rb_vm_t either, and I do think `redefined_flag` fits well with those. But more significantly it seems to result in one fewer instruction. For example:

Before:

```
(lldb) disassemble -n vm_opt_str_freeze
miniruby`vm_exec_core:
miniruby[0x10028233e] <+14558>: movq   0x11a86b(%rip), %rax      ; ruby_current_vm_ptr
miniruby[0x100282345] <+14565>: testb  $0x4, 0x242c(%rax)
```

After:

```
(lldb) disassemble -n vm_opt_str_freeze
ruby`vm_exec_core:
ruby[0x100280ebe] <+14510>: testb  $0x4, 0x120147(%rip)      ; ruby_vm_redefined_flag + 43
```

Introduce BOP_CMP for optimized comparison
---

Prior to this commit the `OPTIMIZED_CMP` macro relied on a method lookup to determine whether `<=>` was overridden. The result of the lookup was cached, but only for the duration of the specific method that initialized the cmp_opt_data cache structure.

With this method lookup, `[x,y].max` is slower than doing `x > y ? x : y` even though there's an optimized instruction for "new array max". (John noticed somebody a proposed micro-optimization based on this fact in https://github.com/mastodon/mastodon/pull/19903.)

```rb
a, b = 1, 2
Benchmark.ips do |bm|
  bm.report('conditional') { a > b ? a : b }
  bm.report('method') { [a, b].max }
  bm.compare!
end
```

Before:

```
Comparison:
         conditional: 22603733.2 i/s
              method: 19820412.7 i/s - 1.14x  (± 0.00) slower
```

This commit replaces the method lookup with a new CMP basic op, which gives the examples above equivalent performance.

After:

```
Comparison:
              method: 24022466.5 i/s
         conditional: 23851094.2 i/s - same-ish: difference falls within error
```

Relevant benchmarks show an improvement to Array#max and Array#min when not using the optimized newarray_max instruction as well. They are noticeably faster for small arrays with the relevant types, and the same or maybe a touch faster on larger arrays.

<details>

<summary>Benchmark output comparing master@5958c305 against these changes</summary>

* array_max_int

  ```
  |                |compare-ruby|built-ruby|
  |:---------------|-----------:|---------:|
  |ary2.max        |     31.546M|   45.045M|
  |                |           -|     1.43x|
  |ary10.max       |     25.907M|   37.594M|
  |                |           -|     1.45x|
  |ary100.max      |     12.092M|   13.123M|
  |                |           -|     1.09x|
  |ary500.max      |      3.775M|    3.964M|
  |                |           -|     1.05x|
  |ary1000.max     |      2.053M|    2.058M|
  |                |           -|     1.00x|
  |ary2000.max     |    713.012k|  813.339k|
  |                |           -|     1.14x|
  |ary3000.max     |    848.320k|  856.825k|
  |                |           -|     1.01x|
  |ary5000.max     |    403.161k|  412.116k|
  |                |           -|     1.02x|
  |ary10000.max    |    208.381k|  209.780k|
  |                |           -|     1.01x|
  |ary20000.max    |    106.438k|  107.331k|
  |                |           -|     1.01x|
  |ary50000.max    |     42.102k|   41.743k|
  |                |       1.01x|         -|
  |ary100000.max   |     21.131k|   21.069k|
  |                |       1.00x|         -|
  |ary1000000.max  |      2.052k|    2.013k|
  |                |       1.02x|         -|
  ```

* array_max_float

  ```
  |               |compare-ruby|built-ruby|
  |:--------------|-----------:|---------:|
  |ary2.max       |     26.882M|   34.843M|
  |               |           -|     1.30x|
  |ary10.max      |     10.299M|   12.453M|
  |               |           -|     1.21x|
  |ary100.max     |      1.321M|    1.286M|
  |               |       1.03x|         -|
  |ary500.max     |    269.818k|  290.909k|
  |               |           -|     1.08x|
  |ary1000.max    |    136.947k|  152.849k|
  |               |           -|     1.12x|
  |ary2000.max    |     56.819k|   57.943k|
  |               |           -|     1.02x|
  |ary3000.max    |     56.480k|   60.751k|
  |               |           -|     1.08x|
  |ary5000.max    |     28.669k|   30.332k|
  |               |           -|     1.06x|
  |ary10000.max   |     14.247k|   15.255k|
  |               |           -|     1.07x|
  |ary20000.max   |      7.037k|    7.691k|
  |               |           -|     1.09x|
  |ary50000.max   |      2.841k|    3.076k|
  |               |           -|     1.08x|
  |ary100000.max  |      1.442k|    1.523k|
  |               |           -|     1.06x|
  ```

* array_max_str

  ```
  |               |compare-ruby|built-ruby|
  |:--------------|-----------:|---------:|
  |ary2.max       |     21.882M|   35.461M|
  |               |           -|     1.62x|
  |ary10.max      |     11.834M|   13.228M|
  |               |           -|     1.12x|
  |ary100.max     |      1.398M|    1.410M|
  |               |           -|     1.01x|
  |ary500.max     |    309.885k|  304.284k|
  |               |       1.02x|         -|
  |ary1000.max    |    138.160k|  138.101k|
  |               |       1.00x|         -|
  |ary2000.max    |     57.870k|   61.428k|
  |               |           -|     1.06x|
  |ary3000.max    |     58.945k|   59.989k|
  |               |           -|     1.02x|
  |ary5000.max    |     30.331k|   31.941k|
  |               |           -|     1.05x|
  |ary10000.max   |     12.037k|   12.395k|
  |               |           -|     1.03x|
  |ary20000.max   |      6.169k|    6.258k|
  |               |           -|     1.01x|
  |ary50000.max   |      2.459k|    2.506k|
  |               |           -|     1.02x|
  |ary100000.max  |      1.023k|    1.035k|
  |               |           -|     1.01x|
  ```

* array_min

  ```
  |                |compare-ruby|built-ruby|
  |:---------------|-----------:|---------:|
  |ary2.min        |     27.778M|   45.249M|
  |                |           -|     1.63x|
  |ary10.min       |     26.667M|   37.313M|
  |                |           -|     1.40x|
  |ary100.min      |     12.346M|   12.970M|
  |                |           -|     1.05x|
  |ary500.min      |      3.413M|    4.088M|
  |                |           -|     1.20x|
  |ary1000.min     |      2.049M|    2.081M|
  |                |           -|     1.02x|
  |ary2000.min     |    847.027k|  823.113k|
  |                |       1.03x|         -|
  |ary3000.min     |    788.830k|  857.045k|
  |                |           -|     1.09x|
  |ary5000.min     |    400.064k|  413.223k|
  |                |           -|     1.03x|
  |ary10000.min    |    205.225k|  209.661k|
  |                |           -|     1.02x|
  |ary20000.min    |    106.158k|  106.014k|
  |                |       1.00x|         -|
  |ary50000.min    |     41.978k|   41.151k|
  |                |       1.02x|         -|
  |ary100000.min   |     20.452k|   20.793k|
  |                |           -|     1.02x|
  |ary1000000.min  |      1.970k|    2.015k|
  |                |           -|     1.02x|
  ```

* array_sort_int

  ```
  |               |compare-ruby|built-ruby|
  |:--------------|-----------:|---------:|
  |ary2.sort      |      4.476M|    4.566M|
  |               |           -|     1.02x|
  |ary10.sort     |      1.883M|    1.952M|
  |               |           -|     1.04x|
  |ary100.sort    |    188.740k|  208.212k|
  |               |           -|     1.10x|
  |ary1000.sort   |     11.633k|   12.404k|
  |               |           -|     1.07x|
  |ary10000.sort  |     870.835|   901.142|
  |               |           -|     1.03x|

* range_min

  ```
  |             |compare-ruby|built-ruby|
  |:------------|-----------:|---------:|
  |(1..10).min  |     35.793M|   45.426M|
  |             |           -|     1.27x|
  ```
  ```
* enum_minmax

  ```
  |                 |compare-ruby|built-ruby|
  |:----------------|-----------:|---------:|
  |set2.min         |      1.966M|    2.267M|
  |                 |           -|     1.15x|
  |set10.min        |      1.340M|    1.479M|
  |                 |           -|     1.10x|
  |set100.min       |    383.083k|  363.135k|
  |                 |       1.05x|         -|
  |set1000.min      |     48.824k|   49.068k|
  |                 |           -|     1.00x|
  |set10000.min     |      5.084k|    5.107k|
  |                 |           -|     1.00x|
  |set2.max         |      1.979M|    2.536M|
  |                 |           -|     1.28x|
  |set10.max        |      1.400M|    1.656M|
  |                 |           -|     1.18x|
  |set100.max       |    381.185k|  368.094k|
  |                 |       1.04x|         -|
  |set1000.max      |     48.235k|   48.075k|
  |                 |       1.00x|         -|
  |set10000.max     |      4.971k|    4.841k|
  |                 |       1.03x|         -|
  |set2.minmax      |      2.278M|    2.352M|
  |                 |           -|     1.03x|
  |set10.minmax     |      1.558M|    1.429M|
  |                 |       1.09x|         -|
  |set100.minmax    |    310.299k|  298.998k|
  |                 |       1.04x|         -|
  |set1000.minmax   |     40.261k|   41.033k|
  |                 |           -|     1.02x|
  |set10000.minmax  |      4.126k|    4.217k|
  |                 |           -|     1.02x|
  ```

* enum_sort

  ```
  |                  |compare-ruby|built-ruby|
  |:-----------------|-----------:|---------:|
  |set2.sort_by      |      1.100M|    1.188M|
  |                  |           -|     1.08x|
  |set10.sort_by     |    558.534k|  608.643k|
  |                  |           -|     1.09x|
  |set100.sort_by    |    107.417k|  116.046k|
  |                  |           -|     1.08x|
  |set1000.sort_by   |     13.077k|   13.804k|
  |                  |           -|     1.06x|
  |set10000.sort_by  |      1.279k|    1.327k|
  |                  |           -|     1.04x|
  ```

</details>