@k0kubun 2022-12-02
Could you benchmark it with either the original motivation, Mastodon, or yjit-bench https://github.com/shopify/yjit-bench? Since this is a type of optimizations that we can't keep adding (if the bitmap width becomes too large, it might need to be more complicated/slower), it's nice to check non-microbenchmarks as well.