@composerinteralia 2022-12-02
Thanks you! I wasn't familiar with https://github.com/shopify/yjit-bench, so I'm glad to have learned about it.

Here's the output I got:

```
master: ruby 3.2.0dev (2022-12-02T01:05:22Z master 8abf9e6ad0) [x86_64-darwin21]
cmp-bop: ruby 3.2.0dev (2022-12-02T02:11:24Z cmp-bop 40f0be0e6a) [x86_64-darwin21]

-------------  -----------  ----------  ------------  ----------  --------------  ---------------
bench          master (ms)  stddev (%)  cmp-bop (ms)  stddev (%)  master/cmp-bop  cmp-bop 1st itr
30k_ifelse     3444.2       3.6         3390.4        1.7         1.02            0.75           
30k_methods    7519.0       1.6         7948.7        8.1         0.95            1.00           
activerecord   279.6        2.0         281.3         1.9         0.99            0.97           
binarytrees    902.7        4.1         850.4         2.1         1.06            1.00           
cfunc_itself   273.8        4.2         259.5         2.0         1.05            1.03           
chunky_png     2943.0       0.2         2857.6        0.4         1.03            1.06           
erubi          450.5        4.7         444.4         2.2         1.01            0.96           
erubi_rails    46.4         4.8         46.1          8.4         1.01            0.97           
etanni         602.2        1.5         548.1         1.2         1.10            1.08           
fannkuchredux  14094.4      7.7         13393.4       0.4         1.05            1.05           
fib            649.5        0.4         564.5         0.3         1.15            0.97           
getivar        330.0        1.9         304.4         0.9         1.08            1.00           
hexapdf        7177.7       1.2         5713.5        1.5         1.26            1.26           
keyword_args   580.4        2.3         511.7         1.5         1.13            1.16           
lee            2482.8       5.5         2438.2        1.0         1.02            1.18           
liquid-render  384.4        11.6        373.4         1.5         1.03            0.96           
mail           239.3        2.5         262.2         1.0         0.91            1.05           
nbody          260.2        5.5         267.1         1.2         0.97            0.93           
optcarrot      16054.5      2.4         15287.1       5.2         1.05            0.99           
psych-load     4896.8       0.3         4499.3        3.6         1.09            1.10           
railsbench     4045.6       4.7         3813.8        3.2         1.06            1.15           
respond_to     481.3        3.7         464.8         6.7         1.04            1.01           
ruby-lsp       419.0        3.1         384.9         5.2         1.09            0.89           
rubykon        24634.5      0.6         23064.7       3.3         1.07            0.96           
setivar        293.7        1.3         302.8         4.3         0.97            0.96           
str_concat     155.3        3.7         167.3         8.0         0.93            1.01           
-------------  -----------  ----------  ------------  ----------  --------------  ---------------
Legend:
- master/cmp-bop: ratio of master/cmp-bop time. Higher is better for cmp-bop. Above 1 represents a speedup.
- cmp-bop 1st itr: ratio of master/cmp-bop time for the first benchmarking iteration.
```

That looks fairly same-ish to me. Do you agree? I also reran a couple individual benchmarks to see how much the results would vary:

```
master: ruby 3.2.0dev (2022-12-02T01:05:22Z master 8abf9e6ad0) [x86_64-darwin21]                  
cmp-bop: ruby 3.2.0dev (2022-12-02T02:11:24Z cmp-bop 40f0be0e6a) [x86_64-darwin21]                
                        
----------  -----------  ----------  ------------  ----------  --------------  ---------------    
bench       master (ms)  stddev (%)  cmp-bop (ms)  stddev (%)  master/cmp-bop  cmp-bop 1st itr     
30k_ifelse  2609.7       1.3         2548.7        1.7         1.02            1.3
hexapdf     5741.8       1.8         5350.2        2.1         1.07            0.99     
mail        241.1        2.1         240.9         1.5         1.00            0.99          
----------  -----------  ----------  ------------  ----------  --------------  ---------------
```