@k0kubun 2022-12-10
@composerinteralia @jhawthorn The first commit of this PR https://github.com/ruby/ruby/commit/c43951e60eed0b01f464cd25441b81751d2d5087 seems to have slowed down MJIT on [optcarrot](https://rubybench.github.io/benchmarks/yjit-bench.html#optcarrot) and [nbody](https://rubybench.github.io/benchmarks/yjit-bench.html#nbody). Here's the comparison of https://github.com/ruby/ruby/commit/c43951e60eed0b01f464cd25441b81751d2d5087 and its parent https://github.com/ruby/ruby/commit/9d4483f24deaf360dafe745a71211ec73dc7029a:

```
before: ruby 3.2.0dev (2022-12-06T14:27:15Z master 9d4483f24d) +MJIT [x86_64-linux]
after: ruby 3.2.0dev (2022-12-06T20:37:23Z master c43951e60e) +MJIT [x86_64-linux]

---------  -----------  ----------  ----------  ----------  ------------  -------------
bench      before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
nbody      38.3         19.4        52.9        2.5         0.72          0.97
optcarrot  1399.2       0.7         1494.7      0.7         0.94          1.00
---------  -----------  ----------  ----------  ----------  ------------  -------------
```

Do you have any idea to make it as fast as before?