@k0kubun 2022-12-10
Out of curiosity, I benchmarked that commit with the interpreter as well. The difference is small, but I can reliably reproduce the interpreter slowdown on nbody:

```
before: ruby 3.2.0dev (2022-12-06T14:27:15Z master 9d4483f24d) [x86_64-linux]
after: ruby 3.2.0dev (2022-12-06T20:37:23Z master c43951e60e) [x86_64-linux]

-----  -----------  ----------  ----------  ----------  ------------  -------------
bench  before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
nbody  67.4         0.4         68.7        0.4         0.98          0.99
-----  -----------  ----------  ----------  ----------  ------------  -------------
```

<details>
<summary>Logs</summary>

```
$ ./run_benchmarks.rb nbody -e "before::/opt/rubies/before/bin/ruby" -e "after::/opt/rubies/after/bin/ruby"
Running benchmark "nbody" (1/1)
setarch x86_64 -R taskset -c 7 /opt/rubies/before/bin/ruby -I ./harness benchmarks/nbody/benchmark.rb
ruby 3.2.0dev (2022-12-06T14:27:15Z master 9d4483f24d) [x86_64-linux]
itr #1: 67ms
itr #2: 67ms
itr #3: 67ms
itr #4: 67ms
itr #5: 67ms
itr #6: 67ms
itr #7: 67ms
itr #8: 67ms
itr #9: 67ms
itr #10: 67ms
itr #11: 67ms
itr #12: 67ms
itr #13: 67ms
itr #14: 67ms
itr #15: 67ms
itr #16: 67ms
itr #17: 67ms
itr #18: 67ms
itr #19: 67ms
itr #20: 67ms
itr #21: 67ms
itr #22: 67ms
itr #23: 67ms
itr #24: 67ms
itr #25: 67ms
itr #26: 67ms
itr #27: 67ms
itr #28: 67ms
itr #29: 67ms
itr #30: 67ms
itr #31: 67ms
itr #32: 67ms
itr #33: 67ms
itr #34: 67ms
itr #35: 67ms
itr #36: 67ms
itr #37: 67ms
itr #38: 67ms
itr #39: 67ms
itr #40: 67ms
itr #41: 67ms
itr #42: 67ms
itr #43: 67ms
itr #44: 67ms
itr #45: 67ms
itr #46: 67ms
itr #47: 67ms
itr #48: 67ms
itr #49: 67ms
itr #50: 67ms
itr #51: 67ms
itr #52: 67ms
itr #53: 67ms
itr #54: 67ms
itr #55: 67ms
itr #56: 67ms
itr #57: 67ms
itr #58: 67ms
itr #59: 68ms
itr #60: 68ms
itr #61: 68ms
itr #62: 67ms
itr #63: 67ms
itr #64: 67ms
itr #65: 67ms
itr #66: 67ms
itr #67: 67ms
itr #68: 67ms
itr #69: 67ms
itr #70: 67ms
itr #71: 67ms
itr #72: 67ms
itr #73: 67ms
itr #74: 67ms
itr #75: 67ms
itr #76: 67ms
itr #77: 67ms
itr #78: 67ms
itr #79: 67ms
itr #80: 67ms
itr #81: 67ms
itr #82: 67ms
itr #83: 67ms
itr #84: 67ms
itr #85: 67ms
itr #86: 67ms
itr #87: 67ms
itr #88: 67ms
itr #89: 67ms
itr #90: 67ms
itr #91: 67ms
itr #92: 67ms
itr #93: 67ms
itr #94: 67ms
itr #95: 67ms
itr #96: 67ms
itr #97: 67ms
itr #98: 67ms
itr #99: 67ms
itr #100: 67ms
itr #101: 67ms
itr #102: 67ms
itr #103: 67ms
itr #104: 67ms
itr #105: 67ms
itr #106: 67ms
itr #107: 67ms
itr #108: 67ms
itr #109: 67ms
itr #110: 67ms
itr #111: 67ms
itr #112: 67ms
itr #113: 67ms
itr #114: 67ms
itr #115: 67ms
itr #116: 67ms
itr #117: 67ms
itr #118: 68ms
itr #119: 68ms
itr #120: 68ms
itr #121: 67ms
itr #122: 67ms
itr #123: 67ms
itr #124: 67ms
itr #125: 67ms
itr #126: 67ms
itr #127: 67ms
itr #128: 67ms
itr #129: 67ms
itr #130: 67ms
itr #131: 67ms
itr #132: 67ms
itr #133: 67ms
itr #134: 67ms
itr #135: 67ms
itr #136: 67ms
itr #137: 67ms
itr #138: 67ms
itr #139: 67ms
itr #140: 67ms
itr #141: 67ms
itr #142: 67ms
itr #143: 67ms
itr #144: 67ms
itr #145: 67ms
itr #146: 67ms
itr #147: 67ms
itr #148: 67ms
itr #149: 67ms
Average of last 134, non-warmup iters: 67ms
Running benchmark "nbody" (1/1)
setarch x86_64 -R taskset -c 7 /opt/rubies/after/bin/ruby -I ./harness benchmarks/nbody/benchmark.rb
ruby 3.2.0dev (2022-12-06T20:37:23Z master c43951e60e) [x86_64-linux]
itr #1: 68ms
itr #2: 68ms
itr #3: 68ms
itr #4: 68ms
itr #5: 68ms
itr #6: 68ms
itr #7: 68ms
itr #8: 68ms
itr #9: 68ms
itr #10: 68ms
itr #11: 68ms
itr #12: 68ms
itr #13: 68ms
itr #14: 68ms
itr #15: 68ms
itr #16: 69ms
itr #17: 68ms
itr #18: 68ms
itr #19: 68ms
itr #20: 68ms
itr #21: 68ms
itr #22: 68ms
itr #23: 68ms
itr #24: 68ms
itr #25: 68ms
itr #26: 68ms
itr #27: 69ms
itr #28: 69ms
itr #29: 68ms
itr #30: 68ms
itr #31: 68ms
itr #32: 68ms
itr #33: 68ms
itr #34: 68ms
itr #35: 68ms
itr #36: 68ms
itr #37: 68ms
itr #38: 68ms
itr #39: 68ms
itr #40: 68ms
itr #41: 68ms
itr #42: 68ms
itr #43: 68ms
itr #44: 68ms
itr #45: 68ms
itr #46: 68ms
itr #47: 68ms
itr #48: 68ms
itr #49: 68ms
itr #50: 68ms
itr #51: 68ms
itr #52: 68ms
itr #53: 68ms
itr #54: 68ms
itr #55: 68ms
itr #56: 68ms
itr #57: 68ms
itr #58: 68ms
itr #59: 68ms
itr #60: 68ms
itr #61: 68ms
itr #62: 68ms
itr #63: 68ms
itr #64: 68ms
itr #65: 68ms
itr #66: 68ms
itr #67: 68ms
itr #68: 68ms
itr #69: 68ms
itr #70: 68ms
itr #71: 68ms
itr #72: 68ms
itr #73: 68ms
itr #74: 68ms
itr #75: 68ms
itr #76: 68ms
itr #77: 68ms
itr #78: 68ms
itr #79: 68ms
itr #80: 68ms
itr #81: 68ms
itr #82: 68ms
itr #83: 68ms
itr #84: 68ms
itr #85: 69ms
itr #86: 69ms
itr #87: 69ms
itr #88: 68ms
itr #89: 68ms
itr #90: 68ms
itr #91: 68ms
itr #92: 68ms
itr #93: 68ms
itr #94: 68ms
itr #95: 68ms
itr #96: 68ms
itr #97: 68ms
itr #98: 68ms
itr #99: 68ms
itr #100: 68ms
itr #101: 68ms
itr #102: 68ms
itr #103: 68ms
itr #104: 68ms
itr #105: 68ms
itr #106: 68ms
itr #107: 68ms
itr #108: 68ms
itr #109: 68ms
itr #110: 68ms
itr #111: 68ms
itr #112: 68ms
itr #113: 68ms
itr #114: 68ms
itr #115: 68ms
itr #116: 68ms
itr #117: 68ms
itr #118: 68ms
itr #119: 68ms
itr #120: 68ms
itr #121: 68ms
itr #122: 68ms
itr #123: 68ms
itr #124: 68ms
itr #125: 68ms
itr #126: 68ms
itr #127: 68ms
itr #128: 68ms
itr #129: 68ms
itr #130: 68ms
itr #131: 68ms
itr #132: 68ms
itr #133: 68ms
itr #134: 68ms
itr #135: 68ms
itr #136: 68ms
itr #137: 68ms
itr #138: 68ms
itr #139: 68ms
itr #140: 68ms
itr #141: 69ms
itr #142: 68ms
itr #143: 69ms
itr #144: 69ms
itr #145: 69ms
itr #146: 68ms
Average of last 131, non-warmup iters: 68ms
Total time spent benchmarking: 20s

before: ruby 3.2.0dev (2022-12-06T14:27:15Z master 9d4483f24d) [x86_64-linux]
after: ruby 3.2.0dev (2022-12-06T20:37:23Z master c43951e60e) [x86_64-linux]

-----  -----------  ----------  ----------  ----------  ------------  -------------
bench  before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
nbody  67.4         0.4         68.7        0.4         0.98          0.99
-----  -----------  ----------  ----------  ----------  ------------  -------------
Legend:
- before/after: ratio of before/after time. Higher is better for after. Above 1 represents a speedup.
- after 1st itr: ratio of before/after time for the first benchmarking iteration.
```

</details>

In the Logs, you can see it used to be consistently 67ms (with a few 68ms) whereas the first commit makes it 68ms (with a few 69ms).